continuous_auth_service/requirements.txt
flask
flask-cors
pymongo
scikit-learn
pandas
numpy
python-dotenv
pyjwt
